Added things back to Notifications branch
